Split GradingSystemComponent as Interval (implemented here) and Detailed (existing).
