+++ src/main/webapp/app/course/manage/course-management.route.ts
@@ -66,6 +66,7 @@ export const courseManagementState: Routes = [
             pageTitle: 'artemisApp.course.gradingSystem',
         },
         canActivate: [UserRouteAccessService],
+        loadChildren: () => import('app/grading-system/grading-system.module').then((m) => m.GradingSystemModule),
     },
     {
         path: ':courseId/plagiarism-cases',
