+++ src/main/webapp/app/grading-system/grading-system.component.scss
@@ -1,23 +1,62 @@
-.dropdown-container {
+@import 'src/main/webapp/app/overview/overview-vars';
+
+.tab-bar {
     display: flex;
-    justify-content: space-between;
-    width: 100%;
-    margin: 1em 0;
-}
+    align-items: center;
+    min-height: 60px;
+    margin: 0 -1.25rem;
+    margin-bottom: 0.5rem;
+    $padding-y: 11px;
+    padding-top: $padding-y;
+    padding-bottom: $padding-y;
+    border-bottom: 1px solid $border-color;
 
-.dropdown-container .form-group {
-    width: 23%;
-    min-width: 20%;
+    .tab-item {
+        cursor: pointer;
+        margin: 0 0.5em;
 
-    .form-control {
-        margin-top: 0.25em;
+        &.active {
+            box-shadow: 0 2px 0 $artemis-dark;
+        }
+    }
+
+    > div:first-child {
+        display: flex;
+        align-items: center;
+        justify-content: space-between;
+        gap: 0.5rem;
+
+        > div:first-child {
+            display: flex;
+            align-items: center;
+        }
+
+        .controls {
+            display: flex;
+            align-items: center;
+            gap: 0.5rem;
+            margin-right: 3px;
+            margin-left: 3px;
+        }
+    }
+
+    > div:first-child.no-indent {
+        justify-content: start;
+
+        .controls {
+            padding-right: 0.5rem;
+            margin-right: 0;
+        }
     }
 }
 
-.grading-scale-table-row td {
-    width: 17%;
+@media screen and (max-width: 991px) {
+    .tab-bar > div:first-child {
+        flex-wrap: wrap;
 
-    input {
-        min-width: 50%;
+        .controls {
+            padding-right: 0.5rem;
+            margin-right: 0;
+        }
     }
 }
