+++ src/main/webapp/app/grading-system/grading-system.module.ts
@@ -1,12 +1,17 @@
 import { NgModule } from '@angular/core';
 import { ArtemisSharedModule } from 'app/shared/shared.module';
 import { DeviceDetectorService } from 'ngx-device-detector';
-import { GradingSystemComponent } from 'app/grading-system/grading-system.component';
+import { DetailedGradingSystemComponent } from 'app/grading-system/detailed-grading-system/detailed-grading-system.component';
 import { GradingSystemInfoModalComponent } from 'app/grading-system/grading-system-info-modal/grading-system-info-modal.component';
+import { GradingSystemComponent } from 'app/grading-system/grading-system.component';
+import { gradingSystemState } from 'app/grading-system/grading-system.route';
+import { RouterModule } from '@angular/router';
+import { IntervalGradingSystemComponent } from 'app/grading-system/interval-grading-system/interval-grading-system.component';
+import { ArtemisModePickerModule } from 'app/exercises/shared/mode-picker/mode-picker.module';
 
 @NgModule({
-    declarations: [GradingSystemComponent, GradingSystemInfoModalComponent],
-    imports: [ArtemisSharedModule],
+    declarations: [GradingSystemComponent, DetailedGradingSystemComponent, IntervalGradingSystemComponent, GradingSystemInfoModalComponent],
+    imports: [ArtemisSharedModule, RouterModule.forChild(gradingSystemState), ArtemisModePickerModule],
     exports: [GradingSystemComponent, GradingSystemInfoModalComponent],
     entryComponents: [GradingSystemComponent],
     providers: [DeviceDetectorService],
