+++ src/main/webapp/app/shared/layouts/navbar/navbar.component.ts
@@ -249,6 +249,8 @@ export class NavbarComponent implements OnInit, OnDestroy {
         tutor_effort_statistics: 'artemisApp.textExercise.tutorEffortStatistics.title',
         text_cluster_statistics: 'artemisApp.textExercise.clusterStatistics.title',
         user_settings: 'artemisApp.userSettings.title',
+        detailed: 'artemisApp.gradingSystem.detailedTab.title',
+        interval: 'artemisApp.gradingSystem.intervalTab.title',
     };
 
     /**
